refactor: util func
